+++ src/tables/trusted_domain.rs
@@ -4,7 +4,7 @@ use anyhow::{bail, Result};
 use redb::ReadableTable;
 use rocksdb::OptimisticTransactionDB;
 
-use super::{KeyValue, UniqueKey};
+use super::{KeyValue, UniqueKey, Value};
 use crate::{types::FromKeyValue, Map, Table};
 
 #[derive(Debug, Clone, PartialEq)]
@@ -70,16 +70,6 @@ impl<'db, 'n, 'd> Table<'db, 'n, 'd, TrustedDomain, &str, &str> {
     ///
     /// Returns an error if the serialization fails or the database operation fails.
     pub fn update(&self, old: &TrustedDomain, new: &TrustedDomain) -> Result<()> {
-<<<<<<< HEAD
-        self.map.update(
-            (old.unique_key(), old.value()),
-            (new.unique_key(), new.value()),
-        )
-||||||| parent of 65d6bb8 (Use redb for trusted domain names)
-        let (ok, ov) = (old.unique_key(), old.value());
-        let (nk, nv) = (new.unique_key(), new.value());
-        self.map.update((ok, &ov), (nk, &nv))
-=======
         let txn = self.db.begin_write()?;
         let mut tbl = txn.open_table::<&str, &str>(self.def)?;
         let Some(val) = tbl.get(old.db_key())? else {
@@ -94,7 +84,6 @@ impl<'db, 'n, 'd> Table<'db, 'n, 'd, TrustedDomain, &str, &str> {
         drop(tbl);
         txn.commit()?;
         Ok(())
->>>>>>> 65d6bb8 (Use redb for trusted domain names)
     }
 
     /// Removes a `TrustedDomain` with the given name.