+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The `Key` and `Value` traits have been introduced to provide a standardized
+  way to interact with key-value pairs in `Table`.
+
 ## [0.32.0] - 2024-11-07
 
 ### Changed
@@ -721,7 +728,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Modified `FtpBruteForce` by adding an `is_internal` field which is a boolean
   indicating whether it is internal or not.
 
+<<<<<<< HEAD
 [0.32.0]: https://github.com/petabi/review-database/compare/0.31.0...0.32.0
+||||||| parent of 77fe449 (Add Key, Value, and KeyValue traits)
+=======
+[Unreleased]: https://github.com/petabi/review-database/compare/0.31.0...main
+>>>>>>> 77fe449 (Add Key, Value, and KeyValue traits)
 [0.31.0]: https://github.com/petabi/review-database/compare/0.30.0...0.31.0
 [0.30.0]: https://github.com/petabi/review-database/compare/0.29.1...0.30.0
 [0.29.1]: https://github.com/petabi/review-database/compare/0.29.0...0.29.1